Make money (#2)
* start adr

Not done obviously, but a good initial draft.

* Start on tests and currency

Money is a money type - it stores money as cents. In other words, $1.23
is 123. Money can be multiplied, and it can be divided as long as the
division works equally. For instance, $0.01 or 1 cannot be divided by
two since you can't have half pennies. There may be some other reasons
we may want to support that in the future but for now? This is the way.